Abstract out Elo calculation
